python-workers/requirements.txt
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
greenlet==3.0.3
